# ipx/docs/swagger.yaml
basePath: /
definitions:
  misc.AllowanceCalldataRequest:
    properties:
      owner:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      spender:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  misc.AllowanceCalldataResponse:
    properties:
      data:
        type: string
    type: object
  misc.ApproveCalldataRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      spender:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  misc.ApproveCalldataResponse:
    properties:
      data:
        type: string
    type: object
  misc.BalanceOfCalldataRequest:
    properties:
      account:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  misc.BalanceOfCalldataResponse:
    properties:
      data:
        type: string
    type: object
...
  misc.DecodeCallRequest:
    properties:
      data:
        example: 0xa9059cbb000000000000000000000000da70aa79f1a329719b9cf9d334b0a82b1d5269f300000000000000000000000000000000000000000000000000000000000003e8
        type: string
      signature:
        example: transfer(address,uint256)
        type: string
    type: object
  misc.DecodeCallResponse:
    properties:
      selector:
        type: string
      values:
        additionalProperties:
          type: string
        type: object
    type: object
  misc.DecodeResultRequest:
    properties:
      data:
        example: "0x000000000000000000000000000000000000000000000000000000000001cf1d"
        type: string
      types:
        example:
        - '["uint256"]'
        items:
          type: string
        type: array
    type: object
  misc.DecodeResultResponse:
    properties:
      values:
        items:
          type: string
        type: array
    type: object
...
  misc.EncodeRequest:
    properties:
      args:
        example:
        - '["0xDa70aA79..."'
        - '"1000000000000000000"]'
        items:
          type: string
        type: array
      signature:
        example: transfer(address,uint256)
        type: string
    type: object
  misc.EncodeResponse:
    properties:
      data:
        type: string
    type: object
...
  misc.SelectorRequest:
    properties:
      signature:
        example: transfer(address,uint256)
        type: string
    type: object
  misc.SelectorResponse:
    properties:
      selector:
        type: string
    type: object
...
  misc.TransferCalldataRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  misc.TransferCalldataResponse:
    properties:
      data:
        type: string
    type: object
  misc.TransferFromCalldataRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  misc.TransferFromCalldataResponse:
    properties:
      data:
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
  /evm/abi/decode/call:
    post:
      consumes:
      - application/json
      description: Decodes calldata (4-byte selector + args) into a name→value map
        using the given function signature. Parameter names are used as keys; falls
        back to "arg0", "arg1", … when the signature has no names.
      parameters:
      - description: Function signature and hex-encoded calldata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.DecodeCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.DecodeCallResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Decode ABI-encoded function calldata
      tags:
      - abi
  /evm/abi/decode/result:
    post:
      consumes:
      - application/json
      description: Decodes the raw hex output of an eth_call into human-readable values
        given a list of ABI types
      parameters:
      - description: ABI types and hex-encoded return data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.DecodeResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.DecodeResultResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Decode ABI-encoded return data
      tags:
      - abi
  /evm/abi/encode:
    post:
      consumes:
      - application/json
      description: |-
        Returns the ABI-encoded calldata (4-byte selector + packed arguments) for the given signature and args.
        Signature must be in canonical form — no spaces, no parameter names (e.g. "transfer(address,uint256)").
        Supported types: address | bool | string | bytes | bytes1–bytes32 | uint8/16/32/64/128/256 | int8/16/32/64/128/256
        Tuple, array, and slice types are not yet supported.
      parameters:
      - description: Function signature and arguments
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.EncodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.EncodeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: ABI-encode a function call
      tags:
      - abi
  /evm/abi/encode/allowance:
    post:
      consumes:
      - application/json
      description: Returns ABI-encoded calldata for allowance(address,address)
      parameters:
      - description: Owner and spender addresses
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.AllowanceCalldataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.AllowanceCalldataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build allowance calldata
      tags:
      - abi
  /evm/abi/encode/approve:
    post:
      consumes:
      - application/json
      description: Returns ABI-encoded calldata for approve(address,uint256)
      parameters:
      - description: Spender address and amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.ApproveCalldataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.ApproveCalldataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build approve calldata
      tags:
      - abi
  /evm/abi/encode/balance-of:
    post:
      consumes:
      - application/json
      description: Returns ABI-encoded calldata for balanceOf(address)
      parameters:
      - description: Account address
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.BalanceOfCalldataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.BalanceOfCalldataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build balanceOf calldata
      tags:
      - abi
  /evm/abi/encode/transfer:
    post:
      consumes:
      - application/json
      description: Returns ABI-encoded calldata for transfer(address,uint256)
      parameters:
      - description: Recipient address and amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.TransferCalldataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.TransferCalldataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build transfer calldata
      tags:
      - abi
  /evm/abi/encode/transfer-from:
    post:
      consumes:
      - application/json
      description: Returns ABI-encoded calldata for transferFrom(address,address,uint256)
      parameters:
      - description: From, to addresses and amount
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.TransferFromCalldataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.TransferFromCalldataResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build transferFrom calldata
      tags:
      - abi
  /evm/abi/selector:
    post:
      consumes:
      - application/json
      description: Returns the 4-byte selector for the given function signature
      parameters:
      - description: Function signature
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/misc.SelectorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/misc.SelectorResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Compute ABI function selector
      tags:
      - abi
...
swagger: "2.0"